backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
python-dotenv
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
requests
